audit_log/requirements.txt
ansible==5.3.0
ansible-core==2.12.2
APScheduler==3.8.1
attrs==21.4.0
boto==2.49.0
boto3==1.21.7
botocore==1.24.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
clickclick==20.10.2
connexion==2.14.1
cryptography==36.0.1
Flask==2.0.2
Flask-Cors==3.0.10
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
kazoo==2.5.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.2
packaging==21.3
protobuf==3.19.4
pycparser==2.21
pycryptodome==3.14.0
pykafka==2.8.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.16.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.27.1
resolvelib==0.5.4
s3transfer==0.5.2
scapy==2.4.5
six==1.16.0
SQLAlchemy==1.4.31
stanfordkarel==0.2.6
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.2
